src/main/java/com/whujl/array/SearchIn2DMatrixII.java
package com.whujl.array;

/**
 * https://leetcode.com/problems/search-a-2d-matrix-ii/
 */
public class SearchIn2DMatrixII {

    public boolean searchMatrix(int[][] matrix, int target) {
        if (matrix == null || matrix.length == 0) {
            return false;
        }

        if (matrix[0] == null || matrix[0].length == 0) {
            return false;
        }

        int n = matrix.length;
        int m = matrix[0].length;
        int x = n - 1;
        int y = 0;

        while (x >= 0 && y < m) {
            if (matrix[x][y] < target) {
                y++;
            } else if (matrix[x][y] > target) {
                x--;
            } else {
                return true;
            }
        }

        return false;
    }
}
